+++ R/tidyHtmlTable_helpers_extractElementsAndConvertToTbl.R
@@ -3,7 +3,7 @@
 #' @param x `list` with columns to be joined
 #' @param elements `char` vector with the elements to select
 prExtractElementsAndConvertToTbl <- function(x, elements) {
-  x[elements] %>%
-    prBindDataListIntoColumns() %>%
+  x[elements] |>
+    prBindDataListIntoColumns() |>
     dplyr::distinct()
 }