+++ inst/examples/interactiveTable_example.R
@@ -1,4 +1,3 @@
-library(magrittr)
 # A simple output
 long_txt <- "Lorem ipsum dolor sit amet, consectetur adipiscing elit,
 sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.
@@ -8,10 +7,12 @@ in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur
 sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt
 mollit anim id est laborum"
 short_txt <- gsub("(^[^.]+).*", "\\1", long_txt)
+tbl_input <- cbind(rep(short_txt, 2), rep(long_txt, 2))
 
-cbind(rep(short_txt, 2),
-      rep(long_txt, 2)) %>%
-  addHtmlTableStyle(col.rgroup = c("#FFF", "#EEF")) %>%
-  interactiveTable(minimized.columns = ncol(.),
-                   header = c("Short", "Long"),
-                   rnames = c("First", "Second"))
+tbl_input |>
+  addHtmlTableStyle(col.rgroup = c("#FFF", "#EEF")) |>
+  interactiveTable(
+    minimized.columns = ncol(tbl_input),
+    header = c("Short", "Long"),
+    rnames = c("First", "Second")
+  )